+++ packages/applications/scheduled-tasks/src/gestionnaireRéseau/addNewGestionnairesDeRéseau.ts
@@ -1,0 +1,42 @@
+import { GestionnaireRéseau } from '@potentiel-domain/reseau';
+import { getLogger } from '@potentiel-libraries/monitoring';
+import { mediator } from 'mediateur';
+import { Params } from '.';
+
+export const addNewGestionnairesDeRéseau = async ({
+  gestionnairesFromORE,
+  gestionnairesRéseau,
+}: Params) => {
+  const gestionnairesRéseauEICFromDb = gestionnairesRéseau.items.map(
+    (gestionnaire) => gestionnaire.identifiantGestionnaireRéseau.codeEIC,
+  );
+
+  const newGestionnairesRéseaux = gestionnairesFromORE.filter(
+    (gestionnaire) => !gestionnairesRéseauEICFromDb.includes(gestionnaire.codeEIC),
+  );
+
+  newGestionnairesRéseaux.length
+    ? getLogger().info(
+        '[updateGestionnaireDeRéseau] Des nouveaux gestionnaires de réseau vont être ajoutés',
+        { nouveauxGestionnaires: newGestionnairesRéseaux },
+      )
+    : getLogger().info(
+        "[updateGestionnaireDeRéseau] Il n'y a pas de nouveaux gestionnaires de réseaux",
+      );
+
+  for (const newGestionnaireRéseaux of newGestionnairesRéseaux) {
+    await mediator.send<GestionnaireRéseau.GestionnaireRéseauUseCase>({
+      type: 'Réseau.Gestionnaire.UseCase.AjouterGestionnaireRéseau',
+      data: {
+        aideSaisieRéférenceDossierRaccordementValue: {
+          expressionReguliereValue: '',
+          formatValue: '',
+          légendeValue: '',
+        },
+        identifiantGestionnaireRéseauValue: newGestionnaireRéseaux.codeEIC,
+        raisonSocialeValue: newGestionnaireRéseaux.raisonSociale,
+        contactEmailValue: newGestionnaireRéseaux.contactEmail,
+      },
+    });
+  }
+};
